requirements.txt
asgiref==3.7.2
autopep8==2.0.4
black==24.2.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.3
Django==3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-twilio==0.8.0
djangorestframework==3.14.0
drf-yasg==1.21.7
firebase-admin==6.4.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
msgpack==1.0.7
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.12.1
phonenumbers==8.13.18
Pillow==9.5.0
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pyotp==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
rave-python==1.4.0
requests==2.31.0
rsa==4.9
setuptools==69.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
twilio==7.17.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.2.0
uuid==1.30
whitenoise==6.5.0
